Rename components/commons to components/common; update imports
