+++ components/prompt-kit/markdown.tsx
@@ -5,7 +5,7 @@ import { memo, useId, useMemo } from "react"
 import ReactMarkdown, { Components } from "react-markdown"
 import remarkBreaks from "remark-breaks"
 import remarkGfm from "remark-gfm"
-import { ButtonCopy } from "@/components/commons/button-copy"
+import { ButtonCopy } from "@/components/common/button-copy"
 import {
   CodeBlock,
   CodeBlockCode,
